+++ packages/terra-clinical-data-grid/src/terra-dev-site/doc/clinical-data-grid/DatagridExamples.2/DatagridWithNoPinnedColumns.2.doc.jsx
@@ -3,7 +3,7 @@ import ExampleTemplate from 'terra-doc-template/lib/ExampleTemplate';
 
 /* eslint-disable import/no-webpack-loader-syntax, import/first, import/extensions, import/no-unresolved, import/no-duplicates */
 import DatagridWithNoPinnedColumns from './DatagridWithNoPinnedColumns';
-import DatagridWithNoPinnedColumnsSrc from '!raw-loader!../../../../../src/terra-dev-site/doc/clinical-data-grid/DatagridExamples/DatagridWithNoPinnedColumns';
+import DatagridWithNoPinnedColumnsSrc from '!raw-loader!../../../../../src/terra-dev-site/doc/clinical-data-grid/DatagridExamples.2/DatagridWithNoPinnedColumns';
 /* eslint-enable import/no-webpack-loader-syntax, import/first, import/extensions, import/no-unresolved, import/no-duplicates */
 
 const DatagridWithNoPinnedColumnsExample = () => (