+++ packages/terra-clinical-detail-view/src/terra-dev-site/doc/clinical-detail-view/clinicalDetailView.1.doc.jsx
@@ -1,21 +1,21 @@
 import React from 'react';
 import DocTemplate from 'terra-doc-template';
-import ReadMe from '../../../docs/README.md';
-import { name } from '../../../package.json';
+import ReadMe from '../../../../docs/README.md';
+import { name } from '../../../../package.json';
 
 /* eslint-disable import/no-webpack-loader-syntax, import/first, import/extensions, import/no-unresolved, import/no-duplicates */
 // Component Source
-import DetailViewSrc from '!raw-loader!../../../src/DetailView';
-import DetailListSrc from '!raw-loader!../../../src/DetailList';
-import DetailListItemSrc from '!raw-loader!../../../src/DetailListItem';
+import DetailViewSrc from '!raw-loader!../../../../src/DetailView';
+import DetailListSrc from '!raw-loader!../../../../src/DetailList';
+import DetailListItemSrc from '!raw-loader!../../../../src/DetailListItem';
 
 // Example files
-import DetailViewDivided from './example/DetailViewDivided';
-import DetailViewDividedSrc from '!raw-loader!../../../src/terra-dev-site/doc/example/DetailViewDivided';
-import DetailViewNoDivider from './example/DetailViewNoDivider';
-import DetailViewNoDividerSrc from '!raw-loader!../../../src/terra-dev-site/doc/example/DetailViewNoDivider';
-import DetailViewDividedSmallerTitles from './example/DetailViewDividedSmallerTitles';
-import DetailViewDividedSmallerTitlesSrc from '!raw-loader!../../../src/terra-dev-site/doc/example/DetailViewDividedSmallerTitles';
+import DetailViewDivided from '../example/DetailViewDivided';
+import DetailViewDividedSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/DetailViewDivided';
+import DetailViewNoDivider from '../example/DetailViewNoDivider';
+import DetailViewNoDividerSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/DetailViewNoDivider';
+import DetailViewDividedSmallerTitles from '../example/DetailViewDividedSmallerTitles';
+import DetailViewDividedSmallerTitlesSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/DetailViewDividedSmallerTitles';
 /* eslint-enable import/no-webpack-loader-syntax, import/first, import/extensions, import/no-unresolved, import/no-duplicates */
 
 const DocPage = () => (