+++ packages/terra-clinical-item-collection/src/terra-dev-site/doc/clinical-item-collection/clinicalItemCollection.1.doc.jsx
@@ -1,16 +1,16 @@
 import React from 'react';
 import DocTemplate from 'terra-doc-template';
-import ReadMe from '../../../docs/README.md';
-import { name } from '../../../package.json';
+import ReadMe from '../../../../docs/README.md';
+import { name } from '../../../../package.json';
 
 /* eslint-disable import/no-webpack-loader-syntax, import/first, import/extensions, import/no-unresolved, import/no-duplicates */
 // Component Source
-import ItemCollectionSrc from '!raw-loader!../../../src/ItemCollection';
-import ItemSrc from '!raw-loader!../../../src/Item';
+import ItemCollectionSrc from '!raw-loader!../../../../src/ItemCollection';
+import ItemSrc from '!raw-loader!../../../../src/Item';
 
 // Example Files
-import ItemCollectionExample from './example/ItemCollectionExample';
-import ItemCollectionExampleSrc from '!raw-loader!./../../../src/terra-dev-site/doc/example/ItemCollectionExample';
+import ItemCollectionExample from '../example/ItemCollectionExample';
+import ItemCollectionExampleSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/ItemCollectionExample';
 /* eslint-enable import/no-webpack-loader-syntax, import/first, import/extensions, import/no-unresolved, import/no-duplicates */
 
 const DocPage = () => (